requirements.txt
streamlit
langchain
google-generativeai
langchain-google-genai
edge-tts
SpeechRecognition
python-dotenv
pdfplumber
plotly
langgraph
reportlab
pdf2image
pytesseract
Pillow
pyresparser
langchain-community
pandas
pyaudio
requests
regex
sounddevice
wavio
numpy